Support GOT being local to the function definition.
Fix problems not defining temporary labels that are exposed in
GOT offsets.

The compiler now generates code that expects AR(0) to contain
the base for a secret section that is used in conjunction with
checking the stack integrity (by offset 196 not being overwritten).

This usage is documeted in the ABI by AR0 being reerved for
system use.

What mindset is this?