Handle %f register specifer just like %g and %a.
Initially there was some idea to magically limit gcc's use of floating
point registers to the four of the old architecture, but that is a
pipe dream.  One could add code to check for the use of AFP and report
so that the end user will not get an 0c6.